+++ go.mod
@@ -1,3 +1,3 @@
-module quamina
+module 	github.com/timbray/quamina
 
 go 1.18
