+++ pkg/pruner/doc.go
