+++ pkg/quamina/arrays_test.go
