+++ pkg/quamina/concurrency_test.go
@@ -39,7 +39,7 @@ func TestConcurrency(t *testing.T) {
 	// I was initially surprised that adding 860 or so changes to the automaton while it's running doesn't seem to
 	//  cause any decrease in performance. But I guess it splits out very cleanly onto another core and really
 	//  doesn't steal any resources from the thread doing the Match calls
-	file, err := os.Open("../test_data/citylots.jlines")
+	file, err := os.Open(testData("citylots.jlines"))
 	if err != nil {
 		t.Error("Can't open file: " + err.Error())
 	}
