+++ pkg/quamina/flattener.go
