+++ pkg/quamina/pattern.go
