+++ pkg/quamina/pattern_test.go
