+++ pkg/quamina/small_table.go
