+++ pkg/quamina/small_table_test.go
